+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-### 5.0.0
+### 5.0.0-rc.3
+Because signals always have a value, when a signal is passed to the effect created by `createEffect()`, its initial value will now be emitted to the effect handler function.  
+Previously, the initial value was only emitted during `effect()` execution.
+
+### 5.0.0-rc.2
 * Angular v20 is now supported;
 * The generator callback now has the second argument, `callbacks`. It contains an object with callbacks `success` and `error` - the observable you provide to `createEffect()` can use them to notify the effect caller;
 * The second argument of `createEffect()` got new fields: